requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==0.6.23
Pillow==10.2.0
pydantic==2.6.1
langchain==0.1.9
langchain-community==0.0.24
langchain-text-splitters==0.0.1
chromadb==0.4.22
sentence-transformers==2.3.1